requirements.txt
# Reproduction Planner Agent - Dependencies

# Core (required)
openai>=1.0.0
pyyaml>=6.0
requests>=2.28.0

# Anthropic provider (optional, for --api-style anthropic)
anthropic>=0.39.0

# Testing
pytest>=7.0.0
